requirements.txt
google-genai
python-dotenv
pyttsx3
SpeechRecognition
pywin32
pyaudio